Refactor(DK-321): 아이콘 컨벤션에 맞춰 이름 카멜케이스로 변경 - 중복사용된 trashCan 아이콘을 모두 한가지 tsx를 사용하도록 수정
