+++ src/pages/CreateQuiz/composite/QuizWriteForm/QuestionForm.tsx
@@ -12,7 +12,7 @@ import { OXQuestionTemplate } from "@/pages/CreateQuiz/composite/QuizWriteForm/O
 import useAutoResizeTextarea from "@/hooks/useAutoResizeTextArea";
 import { useAtom } from "jotai";
 import useUpdateQuizCreationInfo from "@/hooks/useUpdateQuizCreationInfo";
-import deleteIcon from "/assets/svg/quizWriteForm/delete_ellipse.svg";
+import deleteIcon from "/assets/svg/quizWriteForm/deleteEllipse.svg";
 import { AnswerType, QuizQuestionType } from "@/types/QuizType";
 import QuestionTemplateTypeUtilButton from "./QuestionTemplateTypeUtilButton";
 import { OxQuiz } from "@/svg/QuizWriteForm/OXQuiz";
