+++ src/pages/MyPage/composite/MyStudyGroups/MyStudyGroups.tsx
@@ -14,8 +14,8 @@ import { parseQueryParams } from "@/utils/parseQueryParams";
 import { FetchStudyGroupsParams } from "@/types/ParamsType";
 import { NoDataSection } from "@/components/composite/NoDataSection/NoDataSection";
 import StudyGroupButton from "../../components/StudyGroupButton/StudyGroupButton";
-import textBox from "/public/assets/svg/myPage/text-box.svg";
-import pencilUnderline from "/public/assets/svg/myPage/pencil-underline.svg";
+import textBox from "/public/assets/svg/myPage/textBox.svg";
+import pencilUnderline from "/public/assets/svg/myPage/pencilUnderline.svg";
 import { useNavigate } from "react-router-dom";
 import ROUTES from "@/data/routes";
 import { myStudyGroupFilterAtom } from "@/store/filterAtom";
