--- src/svg/QuizWriteForm/Trash.tsx
@@ -1,23 +1,0 @@
-import React from "react";
-import { SVGProps } from "@/types/SVGProps";
-
-export const Trash: React.FC<SVGProps> = ({ alt, ...props }) => {
-  return (
-    <svg
-      width={props.width}
-      height={props.height}
-      viewBox={`0 0 ${props.width} ${props.height}`}
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <title id="svgTitle">{alt}</title>
-      <path
-        d="M4 6.17647H20M9 3H15M10 16.7647V10.4118M14 16.7647V10.4118M15.5 21H8.5C7.39543 21 6.5 20.0519 6.5 18.8824L6.0434 7.27937C6.01973 6.67783 6.47392 6.17647 7.04253 6.17647H16.9575C17.5261 6.17647 17.9803 6.67783 17.9566 7.27937L17.5 18.8824C17.5 20.0519 16.6046 21 15.5 21Z"
-        stroke={props.stroke}
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
